feature(frontend): Release item fetch
Implements: #3
