+++ frontend/src/Pages/Index/Components/FolderListItem.jsx
@@ -1,23 +1,14 @@
-import {FileOutlined} from "@ant-design/icons";
-import {List} from "antd";
+import {FolderOutlined} from "@ant-design/icons";
+import {Button, List} from "antd";
 
-export function FileListItem({file}) {
+export function FolderListItem({folder, setParentFolderId}) {
     return (
-        <List.Item.Meta
-            avatar={<FileOutlined />}
-            title={file.name}
-            description={`Size: ${formatBytes(file.size)}, Type : ${file.fileType}`}
-        />
+        <Button onClick={() => setParentFolderId(folder.id)}>
+            <List.Item.Meta
+                avatar={<FolderOutlined/>}
+                title={folder.name}
+                /*description="Ant Design, a design language for background applications, is refined by Ant UED Team"*/
+            />
+        </Button>
     )
-}
-
-function formatBytes(bytes, decimals = 2) {
-    if (bytes === 0) return '0 B';
-
-    const k = 1024;
-    const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB'];
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-    const value = parseFloat((bytes / Math.pow(k, i)).toFixed(decimals));
-    return `${value} ${sizes[i]}`;
 }
